Added Quantity to Sale Stocks and Added Call Functionality for Admin Information
I added the variable of quantity to the stocks for sale table, as well as fixing any areas that utilizes the information so no errors will be thrown when accessing any pages. Added the information functionality for the admin_dashbaord so when an administrator is logged in they can see all stocks currently existing, as well as any transaction information.